Move sign_and_send to utils
